+++ src/main/java/com/nttdatacasefirst/stockAPI/entity/enums/CouponType.java
@@ -1,16 +1,16 @@
 package com.nttdatacasefirst.stockAPI.entity.enums;
 
 public enum CouponType {
-    DIVIDEND(1, "DIVIDEND"), //kar payi kuponu (temettu)
-    INTEREST(2, "INTEREST"); //pay alma kuponu
-    private final Integer value;
+    DIVIDEND(1L, "DIVIDEND"), //kar payi kuponu (temettu)
+    INTEREST(2L, "INTEREST"); //pay alma kuponu
+    private final Long value;
     private final String label;
 
-    CouponType(Integer value, String label){
+    CouponType(Long value, String label){
         this.value = value;
         this.label = label;
     }
-    public Integer getValue() { return value; }
+    public Long getValue() { return value; }
     public String getLabel() { return label;}
 
 }
