+++ src/main/java/com/nttdatacasefirst/stockAPI/entity/enums/InvestorType.java
@@ -1,18 +1,18 @@
 package com.nttdatacasefirst.stockAPI.entity.enums;
 
 public enum InvestorType {
-    NATURALPERSON(1, "NATURALPERSON"), //Gercek kisi
-    LEGALENTITY(2, "LEGALENTITY"); //Tuzel kisi
+    NATURALPERSON(1L, "NATURALPERSON"), //Gercek kisi
+    LEGALENTITY(2L, "LEGALENTITY"); //Tuzel kisi
 
-    private final Integer value;
+    private final Long value;
     private final String label;
 
-    InvestorType(Integer value, String label){
+    InvestorType(Long value, String label){
         this.value = value;
         this.label = label;
     }
 
-    public Integer getValue(){ return value; }
+    public Long getValue(){ return value; }
     public String getLabel(){ return label; }
 
 }
