+++ src/main/java/com/nttdatacasefirst/stockAPI/entity/enums/ProcessType.java
@@ -1,15 +1,15 @@
 package com.nttdatacasefirst.stockAPI.entity.enums;
 
 public enum ProcessType {
-    STOCK(1, "STOCK"), //Hisse Senedi islemleri
-    DIVIDEND(2, "DIVIDEND"); //Kar Payi islemleri
+    STOCK(1L, "STOCK"), //Hisse Senedi islemleri
+    DIVIDEND(2L, "DIVIDEND"); //Kar Payi islemleri
 
-    private final Integer value;
+    private final Long value;
     private final String label;
-    ProcessType(Integer value, String label){
+    ProcessType(Long value, String label){
         this.label = label;
         this.value = value;
     }
-    public Integer getValue(){return value; }
+    public Long getValue(){return value; }
     public String getLabel(){ return label; }
 }
